fix: TIO-74: rename import file
